# life/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API documentation for the Life Pratics project",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API server"
        }
    ],
    "paths": {
...
        "/planos": {
            "get": {
                "tags": [
                    "Planos de Saúde"
                ],
                "summary": "Exibe uma lista de planos de saúde",
                "operationId": "b714196faae4dadca97b0af5b6cbeb4d",
                "responses": {
                    "200": {
                        "description": "Lista de planos de saúde"
                    }
                }
            },
            "post": {
                "tags": [
                    "Planos de Saúde"
                ],
                "summary": "Armazena um novo plano de saúde",
                "operationId": "27805ffb4509266202c0d3c42ee0632b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "plano_descricao": {
                                        "type": "string",
                                        "example": "Plano Básico"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Plano de saúde criado com sucesso"
                    }
                }
            }
        },
        "/planos/{id}": {
            "get": {
                "tags": [
                    "Planos de Saúde"
                ],
                "summary": "Exibe um plano de saúde específico",
                "operationId": "1912bf15afe4aef96ac8dea9c78336cd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do plano de saúde"
                    }
                }
            },
            "put": {
                "tags": [
                    "Planos de Saúde"
                ],
                "summary": "Atualiza um plano de saúde específico",
                "operationId": "576a60618b7d6366b59936d27e1a90c5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "plano_descricao": {
                                        "type": "string",
                                        "example": "Plano Premium"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Plano de saúde atualizado com sucesso"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Planos de Saúde"
                ],
                "summary": "Remove um plano de saúde específico",
                "operationId": "bb88b39e85d1394e3f5c28e90b764920",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plano de saúde removido com sucesso"
                    }
                }
            }
        },
        "/procedimentos": {
            "get": {
                "tags": [
                    "Procedimentos"
                ],
                "summary": "Exibe uma lista de procedimentos",
                "operationId": "1fe7cd48529a6c5c60af2f7b7f5690c7",
                "responses": {
                    "200": {
                        "description": "Lista de procedimentos"
                    }
                }
            },
            "post": {
                "tags": [
                    "Procedimentos"
                ],
                "summary": "Armazena um novo procedimento",
                "operationId": "17ac86b054479cb91806b98a70ff24eb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "proc_nome": {
                                        "type": "string",
                                        "example": "Consulta Geral"
                                    },
                                    "proc_valor": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 150
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Procedimento criado com sucesso"
                    }
                }
            }
        },
        "/procedimentos/{id}": {
            "get": {
                "tags": [
                    "Procedimentos"
                ],
                "summary": "Exibe um procedimento específico",
                "operationId": "72511a29341ecfa6b3c5c445d5c28974",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do procedimento"
                    }
                }
            },
            "put": {
                "tags": [
                    "Procedimentos"
                ],
                "summary": "Atualiza um procedimento específico",
                "operationId": "0463bb07accc72ca1731d30575b1f9f5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "proc_nome": {
                                        "type": "string",
                                        "example": "Consulta Especializada"
                                    },
                                    "proc_valor": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 200
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Procedimento atualizado com sucesso"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Procedimentos"
                ],
                "summary": "Remove um procedimento específico",
                "operationId": "8cb9286fda9ba5955372abe9752453df",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Procedimento removido com sucesso"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}